refactor to a single class
